+++ src/backend_cairo/pattern.rs
@@ -29,8 +29,6 @@ pub fn apply(
 
     let global_ts = usvg::Transform::from_native(&cr.get_matrix());
     let (sx, sy) = global_ts.get_scale();
-    // Only integer scaling is allowed.
-    let (sx, sy) = (sx.round(), sy.round());
 
     if sx.is_fuzzy_zero() || sy.is_fuzzy_zero() {
         return;