+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This changelog also contains important changes in dependencies.
 
 ### Fixed
 - Object bounding box calculation.
+- Pattern scaling.
 - (usvg) `color` on `use` resolving.
 - (usvg) `offset` attribute resolving inside the `stop` element.
 - (usvg) Ungrouping of groups with non-inheritable attributes.